home/requirements.txt
alembic==1.5.8
amqp==2.6.1
billiard==3.5.0.5
cachelib==0.1.1
celery==4.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
dnspython==1.16.0
eventlet==0.30.2
Flask==1.1.2
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-Session==0.3.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
idna==2.10
importlib-metadata==3.10.1
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==4.6.11
Mako==1.1.4
MarkupSafe==1.1.1
Pillow==8.2.0
pycparser==2.20
pycryptodomex==3.9.4
PyMySQL==1.0.2
pyOpenSSL==19.1.0
python-alipay-sdk==2.1.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
qiniu==7.3.1
redis==3.5.3
requests==2.25.1
ronglian-sms-sdk==1.0.0
six==1.15.0
SQLAlchemy==1.4.7
typing-extensions==3.7.4.3
urllib3==1.26.4
vine==1.3.0
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.1
